# perf(common): optimize `BitmapIter`