@kwannoel 2023-03-29
Before:
```
noelkwan@Noels-MacBook-Pro benches % cargo bench --bench bitmap -- zeros_iter
   Compiling risingwave_common v0.2.0-alpha (/Users/noelkwan/projects/risingwave/src/common)
    Finished bench [optimized] target(s) in 22.16s
     Running benches/bitmap.rs (/Users/noelkwan/projects/risingwave/target/release/deps/bitmap-1b7d7d1feebf7e97)
zeros_iter              time:   [5.9332 ms 5.9379 ms 5.9430 ms]
                        change: [-0.5865% -0.4578% -0.3226%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 15 outliers among 100 measurements (15.00%)
  11 (11.00%) high mild
  4 (4.00%) high severe


```

After:
```
noelkwan@Noels-MacBook-Pro benches % cargo bench --bench bitmap -- zeros_iter                        
   Compiling risingwave_common v0.2.0-alpha (/Users/noelkwan/projects/risingwave/src/common)
    Finished bench [optimized] target(s) in 22.99s
     Running benches/bitmap.rs (/Users/noelkwan/projects/risingwave/target/release/deps/bitmap-1b7d7d1feebf7e97)
zeros_iter              time:   [4.9628 ms 4.9749 ms 4.9874 ms]
                        change: [+0.7150% +1.0190% +1.3460%] (p = 0.00 < 0.05)
                        Change within noise threshold.
```